Refactored webserver to split c/h
